+++ CHANGELOG.md
@@ -4,6 +4,11 @@
 - Implement appender
 - Tests break with ffigen 14.0.0.  Has a different enum treatment. 
 
+
+## 2024-09-15 (0.3.4)
+- Introduce `fetchRows` to map the result of a DuckDB query to a Dart class.
+
+
 ## 2024-09-15 (0.3.3)
 - Fix bug in `fetch` related to data chunks.  Only the last chunk was actually returned!